Move EffectComponentGroups to subcontrollers and move updateEffect to audio player
