+++ src/main/java/sh/ball/gui/controller/SubController.java
@@ -6,6 +6,7 @@ import javafx.scene.shape.SVGPath;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
+import sh.ball.gui.components.EffectComponentGroup;
 
 import java.util.List;
 import java.util.Map;
@@ -19,6 +20,8 @@ public interface SubController {
   List<Slider> sliders();
 
   List<String> labels();
+
+  List<EffectComponentGroup> effects();
 
   List<? extends Node> save(Document document);
